third_party/WebKit/LayoutTests/fast/dom/shadow/focus-controller-recursion-crash.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div>This test passes if it does not crash.</div>
<input id="first">
<select></select>
<input>
<script>
var sel = document.querySelector('select');
for (var i = 0; i < 50000; ++i) {
   var o = document.createElement('option');
   sel.appendChild(o);
}

document.getElementById('first').focus();

eventSender.keyDown('\t');
eventSender.keyDown('\t');
eventSender.keyDown('\t', ['shiftKey']);
eventSender.keyDown('\t', ['shiftKey']);

shouldBeEqualToString('document.activeElement.id', 'first');
</script>
